Feature/statistics (#179)
* fix: 500 error due to redis cloud

* fix: change from redis cloud to in-mem cached

* feat(statistics): implement GetTop3Product in service

* feat(statistics): implement Statistics Controller

* chore: add category schema and  master_data service

* docs(MasterDataController): add xml document for response type

* docs(Controllers): refactor docs for more clear requirements

* fix(application.csproj): fix conflict

* fix: conflict namespace

+++ Backend/src/ShoeStore.Application/DTOs/CheckOutDTOs/CheckOutItemDto.cs
@@ -4,7 +4,7 @@ public sealed record CheckOutItemDto(
     Guid VariantId,
     string ProductName,
     string? ColorName,
-    int Size,
+    decimal Size,
     decimal UnitPrice,
     int Quantity,
     int StockAvailable,
