+++ Backend/src/ShoeStore.Application/DTOs/InvoiceDetailDTOs/InvoiceDetailResponseDto.cs
@@ -4,7 +4,7 @@ namespace ShoeStore.Application.DTOs.InvoiceDetailDTOs
     public class InvoiceDetailResponseDto
     {
         public string ProductName { get; set; } =  string.Empty;
-        public int Size { get; set; } = 0;
+        public decimal Size { get; set; } = 0;
         public string Color { get; set; } = string.Empty;
         public int Quantity { get; set; } = 0;
         public decimal UnitPrice { get; set; } = 0;
