+++ Backend/src/ShoeStore.Application/DTOs/ProductVariantDTOs/ProductVariantResponeDto.cs
@@ -4,7 +4,7 @@ public class ProductVariantResponseDto
 {
     public Guid PublicId { get; set; }
     public int SizeId { get; set; }
-    public int? Size { get; set; }
+    public decimal? Size { get; set; }
     public int? ColorId { get; set; }
     public string? ColorName { get; set; }
     public int Stock { get; set; }
